## Essential Elements of an Effective PR Description Checklist
- [x] The purpose of the PR, such as "Fix some issue (link existing issues this PR will resolve)".
- [x] The test plan, such as providing test command.
- [x] The test results, such as pasting the results comparison before and after, or e2e results
- [ ] (Optional) The necessary documentation update, such as updating `supported_models.md` and `examples` for a new model.

## Purpose
Swap AB in cutlass moe grouped gemm will reduce padding and reduce latency for M <= 64, because the requirement on tile_m is multiple of 64 while for tile_n it's multiple of 16.  

ps. this PR tunes M <= 64 with swap ab case. I'll have a follow up PR to further tune larger Ms which I found further room for improvement.

## Test
1. Quality test - using existing `test_cutlass_moe.py` which compares Triton fusedmoe kernel. Additionally I added odd M test cases.
```
tests/kernels/moe/test_cutlass_moe.py  >>> all 462 passed 
```
2. Performance Benchmark 

- Models:
  - Qwen3-30B: Qwen/Qwen3-30B-A3B (topk=8)
  - Qwen3-235B: Qwen/Qwen3-235B-A22B (topk=8)
  - Mixtral: RedHatAI/Mixtral-8x7B-Instruct-v0.1-FP8 (topk=2)

Note: bold values are swap_ab, for others mainly to show no regression.

| Batch Size | Qwen3-235B Old (μs) | Qwen3-235B New (μs) | Qwen3-30B Old (μs) | Qwen3-30B New (μs) | Mixtral Old (μs) | Mixtral New (μs) |
|------------|---------------------------|-------------------------|---------------------------|-------------------------|---------------------------|-------------------------|
| 1          | **111.29**                | **98.36** (-11.6%)      | **71.02**                 | **65.27** (-8.1%)       | **67.00**                 | **63.34** (-5.5%)       |
| 2          | **153.77**                | **131.63** (-14.4%)     | **89.72**                 | **78.86** (-12.1%)      | **78.37**                 | **73.49** (-6.2%)       |
| 4          | **197.92**                | **193.19** (-2.4%)      | **98.94**                 | **95.91** (-3.1%)       | **89.08**                 | **84.30** (-5.4%)       |
| 8          | **298.71**                | **291.35** (-2.5%)      | **126.96**                | **123.37** (-2.8%)      | **101.27**                | **92.79** (-8.4%)       |
| 16         | 434.96                    | 436.24                  | 161.07                    | 160.85                  | **102.18**                | **98.33** (-3.8%)       |
| 32         | 578.07                    | 578.11                  | 197.17                    | 197.24                  | **103.29**                | **99.04** (-4.1%)       |
| 512        | 945.65                    | 944.97                  | 358.46                    | 359.01                  | 199.95                    | 199.84                  |
| 4096       | 3853.04                   | 3852.27                 | 1603.69                   | 1600.91                 | 1005.34                   | 1003.64                 |
| 10240      | 9398.15                   | 9390.77                 | 3803.25