@shixianc 2025-07-18
@mgoin I took a look at the error msgs [distributed-tests-4-gpus](https://buildkite.com/vllm/ci/builds/24255#01981a22-1af4-4339-a722-1f3972f42e26) and doesn't seem to be related to my change. 
```
port: 51798, useIpv6: false, code: -98, name: EADDRINUSE, message: address already in use
```
might need your help for a force merge.